Check for a temporary directory
By default a temporary directory is set to `tmp/` using $TMP_DIRECTORY
variable. If that directory doesn't exist, warn user and exit.

The script could create that directory, however, it'd need to have write
permissions to the parent directory (`.') and this can't be always
possible. So instead of writting another code to verify write
permissions, it just warns that a temporary directory is missing.